requirements.txt
aiogram==2.16
corpy==0.4.1
#faiss
faiss_cpu==1.7.2
gensim==4.2.0
numpy==1.19.5
pandas==1.3.5
python-dotenv==0.20.0
sentence_transformers==2.2.0
torch==1.10.0
transformers==4.12.3
ufal.morphodita==1.10.1.1
wget==3.2
protobuf == 3.20
